refactor: 컨트롤러 도메인 구조 재정리 및 파일명 개선
