+++ src/main/java/com/comma/soomteum/domain/external/tourapi/dto/KorService2Response.java
@@ -1,4 +1,4 @@
-package com.comma.soomteum.domain.place.dto;
+﻿package com.comma.soomteum.domain.external.tourapi.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
