+++ src/main/java/com/comma/soomteum/domain/external/tourapi/dto/TourApiErrorResponse.java
@@ -1,4 +1,4 @@
-package com.comma.soomteum.domain.place.dto;
+﻿package com.comma.soomteum.domain.external.tourapi.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
