+++ src/main/java/com/comma/soomteum/domain/external/tourapi/service/KorAreaService.java
@@ -1,12 +1,12 @@
-package com.comma.soomteum.domain.place.service;
+﻿package com.comma.soomteum.domain.external.tourapi.service;
 
-import com.comma.soomteum.domain.place.dto.KorService2Response;
-import com.comma.soomteum.domain.place.dto.TourApiRequestDto;
+import com.comma.soomteum.domain.external.tourapi.dto.KorService2Response;
+import com.comma.soomteum.domain.external.tourapi.dto.TourApiRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Mono;
 
-import static com.comma.soomteum.domain.place.service.KorApiCaller.qpIfPresent;
+import static com.comma.soomteum.domain.external.tourapi.service.KorApiCaller.qpIfPresent;
 
 @Service
 @RequiredArgsConstructor
