+++ src/main/java/com/comma/soomteum/domain/external/tourapi/service/KorDetailService.java
@@ -1,7 +1,7 @@
-package com.comma.soomteum.domain.place.service;
+﻿package com.comma.soomteum.domain.external.tourapi.service;
 
-import com.comma.soomteum.domain.place.dto.response.PlaceDetailResponseDto;
-import com.comma.soomteum.domain.place.dto.TourApiRequestDto;
+import com.comma.soomteum.domain.external.tourapi.dto.response.PlaceDetailResponseDto;
+import com.comma.soomteum.domain.external.tourapi.dto.TourApiRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Mono;
