+++ src/main/java/com/comma/soomteum/domain/place/controller/PlaceController.java
@@ -1,4 +1,4 @@
-package com.comma.soomteum.domain.place.controller;
+﻿package com.comma.soomteum.domain.place.controller;
 
 import com.comma.soomteum.domain.place.dto.response.PlaceDetailIntegratedResponseDto;
 import com.comma.soomteum.domain.place.service.PlaceDetailIntegratedService;
@@ -17,7 +17,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import reactor.core.publisher.Mono;
 
-@Tag(name = "여행지", description = "여행지와 관련된 로직")
+@Tag(name = "여행지 상세", description = "여행지 통합 상세 조회 전용 API")
 @RestController
 @RequestMapping(path = "/api/places", produces = MediaType.APPLICATION_JSON_VALUE)
 @RequiredArgsConstructor
@@ -26,34 +26,22 @@ public class PlaceController {
     private final PlaceDetailIntegratedService placeDetailIntegratedService;
 
     @Operation(
-            summary = "여행지 상세 통합 정보 조회",
-            description = """
-                    여행지 상세 조회 시 필요한 모든 정보를 통합하여 제공합니다.
-                    
-                    **제공 정보:**
-                    - 여행지 이름, 사진, 주소, 지역, 테마, 소개
-                    - 한적함 등급 (1-5단계, -1: 데이터 없음)
-                    - 좋아요 수
-                    - AI 꿀팁 요약 (강릉시만 제공)
-                    - 근처 공영주차장 정보 (강릉시만 제공)
-                    
-                    **강릉시 전용 기능:**
-                    - AI가 생성한 여행 꿀팁 요약
-                    - 주변 5km 이내 공영주차장 최대 5개 정보
-                    """
+            summary = "여행지 통합 상세 조회",
+            description = "여행지 상세 화면에서 필요한 정보를 한 번에 제공합니다. "
+                    + "기본 프로필(이름, 사진, 주소, 지역, 테마, 소개), 한적함 등급, 좋아요 수, AI 꿀팁(강릉시), 근처 공영주차장(강릉시)까지 통합 조회합니다."
     )
     @ApiResponses({
             @io.swagger.v3.oas.annotations.responses.ApiResponse(
-                    responseCode = "200", 
+                    responseCode = "200",
                     description = "통합 정보 조회 성공",
                     content = @Content(schema = @Schema(implementation = PlaceDetailIntegratedResponseDto.class))
             ),
             @io.swagger.v3.oas.annotations.responses.ApiResponse(
-                    responseCode = "404", 
+                    responseCode = "404",
                     description = "여행지를 찾을 수 없음"
             ),
             @io.swagger.v3.oas.annotations.responses.ApiResponse(
-                    responseCode = "500", 
+                    responseCode = "500",
                     description = "서버 내부 오류"
             )
     })
@@ -61,7 +49,7 @@ public class PlaceController {
     public Mono<ApiResponse<PlaceDetailIntegratedResponseDto>> getIntegratedPlaceDetail(
             @Parameter(description = "공공데이터 API의 컨텐츠 ID", required = true, example = "128758")
             @PathVariable String contentId) {
-        
+
         return placeDetailIntegratedService.getIntegratedPlaceDetail(contentId)
                 .map(ApiResponse::ok)
                 .onErrorReturn(new ApiResponse<>(null, false, null, null));
